Added the stack space module
This module takes one type of data parameter across all stations in all 22 years and outputs it as the data frame and .csv file of the mean value across all stations for each year. It also gives data frame and .csv files of the man, min, mean, etc values of the station data for each year.