Stack space output csv file
In this folder are 16 outputs .csv files for each of the 8 different parameters we were trying to analyze. For each parameter, there are two types of files, the df_stack_space and the df_yearlyParam. The df_stack_space is stack(average) all the station data for each year, so there are 22 columns, and each row represents every 10 minutes of input data. The df_yearlyParam is the statistical outputs, like min, max, mean, etc, of each year's data, which is from each column of the df_stack_space file. Those two files help us to find potential correlations between the extrusion rate and the spatial seismic data.